fix: use uint16 for ports

+++ pkg/kubernetes/resources/ingress/adapt.go
@@ -15,7 +15,7 @@ func AdaptFuncToEnsure(
 	host,
 	prefix,
 	service string,
-	servicePort int,
+	servicePort uint16,
 	annotations map[string]string,
 ) (resources.QueryFunc, error) {
